+++ src/components/mdxComponents.module.css
@@ -10,27 +10,29 @@
   cursor: zoom-in;
 }
 
+.heading {
+  padding-top: 2rem;
+  font-weight: 700;
+}
+
+.heading:focus {
+  outline: none;
+}
+
 .h1 {
   font-size: 1.7rem;
   border-bottom: 1px solid var(--content-border);
   padding-bottom: 0.2rem;
-  margin: 4rem 0 1rem;
+  margin-top: 2rem;
 }
 
 .h2 {
   font-size: 1.4rem;
-  margin: 3rem 0 1rem;
+  margin-top: 1rem;
 }
 
 .h3 {
   font-size: 1.1rem;
-  margin: 2rem 0 1rem;
-}
-
-.h1,
-.h2,
-.h3 {
-  font-weight: 700;
 }
 
 .ul {
@@ -49,4 +51,50 @@
 
 .ol {
   margin: 1rem 0;
+}
+
+/* Code block */
+.codeBlock {
+  position: relative;
+  font-size: 0.875rem;
+  font-family: Menlo, Monaco, 'Courier New', monospace;
+}
+
+.fileName {
+  background-color: var(--content-bg-primary);
+  color: var(--text-secondary);
+  position: absolute;
+  top: 0;
+  left: 0;
+  padding: 0.25rem 0.5rem;
+  border-radius: 4px 0 4px 0;
+  font-size: 0.8rem;
+  line-height: 1;
+}
+
+.pre {
+  text-align: left;
+  margin: 1rem auto;
+  padding: 2rem 1rem 1rem;
+  overflow: auto;
+
+  border-radius: 6px 4px 4px 4px;
+  font-family: inherit;
+  line-height: 1.9;
+}
+
+.line {
+  display: table-row;
+}
+
+.lineNo {
+  display: table-cell;
+  text-align: right;
+  padding-right: 1em;
+  user-select: none;
+  opacity: 0.5;
+}
+
+.lineContent {
+  display: table-cell;
 }
