+++ src/components/mdxComponents.tsx
@@ -1,9 +1,50 @@
 import { useImageOverlay } from '@/providers/imageOverlayProvider'
 import { Components } from '@mdx-js/react'
 import Image from 'next/image'
+import Highlight, { defaultProps, Language } from 'prism-react-renderer'
+import theme from 'prism-react-renderer/themes/vsDark'
 import { Fragment } from 'react'
-import { CodeBlock } from './codeBlock'
 import styles from './mdxComponents.module.css'
+
+const CodeBlock: React.VFC<{
+  children: string
+  className: string
+}> = ({ children, className }) => {
+  const splited = className.replace(/language-/, '').split(':')
+  const language = splited[0] as Language
+  const fileName = splited[splited.length - 1]
+
+  return (
+    <div className={styles.codeBlock}>
+      <span className={styles.fileName}>{fileName}</span>
+      <Highlight
+        {...defaultProps}
+        theme={theme}
+        code={children.trim()}
+        language={language}
+      >
+        {({ className, style, tokens, getLineProps, getTokenProps }) => (
+          <pre className={`${className} ${styles.pre}`} style={style}>
+            {tokens.map((line, i) => (
+              <div
+                key={i}
+                {...getLineProps({ line, key: i })}
+                className={styles.line}
+              >
+                <span className={styles.lineNo}>{i + 1}</span>
+                <span className={styles.lineContent}>
+                  {line.map((token, key) => (
+                    <span key={key} {...getTokenProps({ token, key })} />
+                  ))}
+                </span>
+              </div>
+            ))}
+          </pre>
+        )}
+      </Highlight>
+    </div>
+  )
+}
 
 export const MDXComponents: Components = {
   img: (props) => {
@@ -34,21 +75,21 @@ export const MDXComponents: Components = {
   p: (props) => <p className={styles.p}>{props.children}</p>,
   h1: (props) => {
     return (
-      <h1 id={props.children} className={styles.h1}>
+      <h1 id={props.children} className={`${styles.heading} ${styles.h1}`}>
         {props.children}
       </h1>
     )
   },
   h2: (props) => {
     return (
-      <h2 id={props.children} className={styles.h2}>
+      <h2 id={props.children} className={`${styles.heading} ${styles.h2}`}>
         {props.children}
       </h2>
     )
   },
   h3: (props) => {
     return (
-      <h3 id={props.children} className={styles.h3}>
+      <h3 id={props.children} className={`${styles.heading} ${styles.h3}`}>
         {props.children}
       </h3>
     )
